quaint refactor
