+++ markout/src/main/kotlin/io/koalaql/markout/ActionableFiles.kt
@@ -1,0 +1,42 @@
+package io.koalaql.markout
+
+import io.koalaql.markout.files.FileAction
+import java.nio.file.Path
+
+class ActionableFiles(
+    private val paths: Map<Path, FileAction>
+) {
+    fun perform() {
+        paths.forEach { (path, action) -> action.perform(path) }
+    }
+
+    fun expect(): List<Diff> {
+        val visited = linkedMapOf<Path, Boolean>()
+
+        val diffs = arrayListOf<Diff>()
+
+        /* we don't check paths if their parents have already failed */
+        fun visit(path: Path?): Boolean {
+            if (path == null) return true /* handle null parent case */
+            visited[path]?.let { return it }
+
+            val parentExpected = visit(path.parent)
+
+            if (!parentExpected) {
+                visited[path] = false
+                return false
+            }
+
+            val result = paths[path]
+                ?.expect(path, diffs)
+                ?:true
+
+            visited[path] = result
+            return result
+        }
+
+        paths.keys.forEach { visit(it) }
+
+        return diffs
+    }
+}
